project

+++ src/main/java/ra/project_api/service/impl/CategoryServiceImpl.java
@@ -2,10 +2,12 @@ package ra.project_api.service.impl;
 
 ;
 
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import ra.project_api.dto.request.CategoryDTO;
 import ra.project_api.model.Category;
 import ra.project_api.repository.CategoryRepository;
 import ra.project_api.service.CategoryService;
@@ -18,6 +20,8 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Autowired
     private CategoryRepository categoryRepository;
+    @Autowired
+    private ModelMapper modelMapper;
 
     @Override
     public List<Category> findAll() {
@@ -36,14 +40,20 @@ public class CategoryServiceImpl implements CategoryService {
     }
     @Override
     public void deleteCategoryById(Long categoryId) {
+        if (categoryRepository.existsProductByCategoryId(categoryId)) {
+            throw new IllegalStateException("Cannot delete category because it has associated products");
+        }
         categoryRepository.deleteById(categoryId);
     }
-
     @Override
-    public Category update(Category category) {
-         categoryRepository.findById(category.getCategoryId()).orElseThrow(() -> new NoSuchElementException("Sửa khong thành công vì ko tồn tại id đó"));
-    return  categoryRepository.save(category);
+    public Category update(Long categoryId, CategoryDTO categoryDTO) {
+        Category existingCategory = categoryRepository.findById(categoryId)
+                .orElseThrow(() -> new NoSuchElementException("Update failed because the ID does not exist"));
+
+        modelMapper.map(categoryDTO, existingCategory);
+        return categoryRepository.save(existingCategory);
     }
+
 
     @Override
     public Category save(Category category) {
